# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ContentTower | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
  "paths": {
    "/Download/download/{cid}": {
      "get": {
        "tags": [
          "Download"
        ],
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/Stream"
                }
              }
            }
          }
        }
      }
    },
    "/Download/metadata/{cid}": {
      "get": {
        "tags": [
          "Download"
        ],
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FileMetadata"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileMetadata"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileMetadata"
                }
              }
            }
          }
        }
      }
    },
    "/Download/check/{cid}": {
      "get": {
        "tags": [
          "Download"
        ],
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/Quota": {
      "get": {
        "tags": [
          "Quota"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Quota/config": {
      "get": {
        "tags": [
          "Quota"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsView"
                }
              }
            }
          }
        }
      }
    },
    "/Upload": {
      "post": {
        "tags": [
          "Upload"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Cid": {
        "required": [
          "hash"
        ],
        "type": "object",
        "properties": {
          "hash": {
            "type": "string"
          }
        }
      },
      "FileMetadata": {
        "type": "object",
        "properties": {
          "cid": {
            "$ref": "#/components/schemas/Cid"
          },
          "name": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "length": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "storeType": {
            "$ref": "#/components/schemas/StoreRequestType"
          },
          "uploadUtc": {
            "type": "string",
            "format": "date-time"
          },
          "lastActivityUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OptionsView": {
        "type": "object",
        "properties": {
          "cleanupIntervalSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "storeDurationDefaultNominalSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "storeDurationDefaultPressureSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "storeDurationTemporaryNominalSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "storeDurationTemporaryPressureSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "QuotaResponse": {
        "type": "object",
        "properties": {
          "quota": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "used": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "state": {
            "$ref": "#/components/schemas/QuotaState"
          }
        }
      },
      "QuotaState": {
        "type": "integer"
      },
      "StoreRequestType": {
        "enum": [
          "Default",
          "TemporaryFile",
          "PermanentFile"
        ]
      },
      "Stream": {
        "type": "string",
        "format": "binary"
      },
      "UploadRequest": {
        "type": "object",
        "properties": {
          "storeType": {
            "$ref": "#/components/schemas/StoreRequestType"
          },
          "name": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "data": {
            "type": "string",
            "format": "byte"
          }
        }
      },
      "UploadResponse": {
        "type": "object",
        "properties": {
          "contentId": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Download"
    },
    {
      "name": "Quota"
    },
    {
      "name": "Upload"
    }
  ]
}